Accurate pixel/meter ratio 100 meters = 1000 pixels

+++ PIF/src/pif/points/Point3D.java
@@ -3,9 +3,9 @@ package pif.points;
 
 public class Point3D {
 	
-	private int x = 0;
-	private int y = 0;
-	private int z = 0;
+	private double x = 0;
+	private double y = 0;
+	private double z = 0;
 	
 	
 	
@@ -16,7 +16,7 @@ public class Point3D {
 	}
 	
 	
-	public Point3D (int x, int y, int z) {
+	public Point3D (double x, double y, double z) {
 		
 		this.x = x;
 		this.y = y;
@@ -37,32 +37,32 @@ public class Point3D {
 	}
 
 
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 
 
-	public int getZ() {
+	public double getZ() {
 		return z;
 	}
 
 
-	public void setZ(int z) {
+	public void setZ(double z) {
 		this.z = z;
 	}
